Add Module
Moved Frame advantage code to separate module
